requirements.txt
django
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
channels
openai